test: split frontend tests
